+++ native/src/main/java/io/ballerina/stdlib/kafka/utils/KafkaUtils.java
@@ -61,6 +61,8 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.Properties;
 
+import static io.ballerina.runtime.api.TypeTags.ARRAY_TAG;
+import static io.ballerina.runtime.api.TypeTags.OBJECT_TYPE_TAG;
 import static io.ballerina.stdlib.kafka.utils.KafkaConstants.KAFKA_ERROR;
 import static io.ballerina.stdlib.kafka.utils.KafkaConstants.NATIVE_CONSUMER;
 import static io.ballerina.stdlib.kafka.utils.KafkaConstants.NATIVE_CONSUMER_CONFIG;
@@ -73,27 +75,35 @@ public class KafkaUtils {
     private KafkaUtils() {
     }
 
-    public static Object[] getResourceParameters(BObject service, BObject listener,
-                                                 ConsumerRecords records) {
-
-        BArray consumerRecordsArray = ValueCreator.createArrayValue(
-                TypeCreator.createArrayType(getConsumerRecord().getType()));
+    public static Object[] getResourceParameters(BObject service, BObject listener, ConsumerRecords records) {
         String keyType = KafkaConstants.DEFAULT_SER_DES_TYPE;
         String valueType = KafkaConstants.DEFAULT_SER_DES_TYPE;
 
-        if (service.getType().getMethods()[0].getParameterTypes().length == 2) {
-            for (Object record : records) {
-                BMap<BString, Object> consumerRecord = populateConsumerRecord(
-                        (ConsumerRecord) record, keyType, valueType);
-                consumerRecordsArray.append(consumerRecord);
-            }
+        BArray consumerRecordsArray = ValueCreator.createArrayValue(
+                TypeCreator.createArrayType(getConsumerRecord().getType()));
+        for (Object record : records) {
+            BMap<BString, Object> consumerRecord = populateConsumerRecord(
+                    (ConsumerRecord) record, keyType, valueType);
+            consumerRecordsArray.append(consumerRecord);
+        }
+
+        Type[] parameterTypes = service.getType().getMethods()[0].getParameterTypes();
+        if (parameterTypes.length == 2) {
             BObject caller =
                     ValueCreator.createObjectValue(ModuleUtils.getModule(), KafkaConstants.CALLER_STRUCT_NAME);
             KafkaConsumer consumer = (KafkaConsumer) listener.getNativeData(NATIVE_CONSUMER);
             Properties consumerProperties = (Properties) listener.getNativeData(NATIVE_CONSUMER_CONFIG);
             caller.addNativeData(NATIVE_CONSUMER, consumer);
             caller.addNativeData(NATIVE_CONSUMER_CONFIG, consumerProperties);
-            return new Object[]{caller, true, consumerRecordsArray, true};
+            if (parameterTypes[0].getTag() == OBJECT_TYPE_TAG && parameterTypes[1].getTag() == ARRAY_TAG) {
+                return new Object[]{caller, true, consumerRecordsArray, true};
+            } else if (parameterTypes[0].getTag() == ARRAY_TAG && parameterTypes[1].getTag() == OBJECT_TYPE_TAG) {
+                return new Object[]{consumerRecordsArray, true, caller, true};
+            } else {
+                throw KafkaUtils.createKafkaError("Invalid remote function signature");
+            }
+        } else if (parameterTypes.length == 1 && parameterTypes[0].getTag() == ARRAY_TAG) {
+            return new Object[]{consumerRecordsArray, true};
         } else {
             throw KafkaUtils.createKafkaError("Invalid remote function signature");
         }
@@ -832,19 +842,14 @@ public class KafkaUtils {
         }
     }
 
-    public static Type getAttachedFunctionReturnType(BObject serviceObject, String functionName, int paramCount) {
+    public static Type getAttachedFunctionReturnType(BObject serviceObject, String functionName) {
         MethodType function = null;
         MethodType[] resourceFunctions = serviceObject.getType().getMethods();
         for (MethodType resourceFunction : resourceFunctions) {
             if (functionName.equals(resourceFunction.getName())) {
                 function = resourceFunction;
-                Type[] parameterTypes = function.getParameterTypes();
-                if (parameterTypes.length == paramCount) {
-                    Type returnType = function.getReturnType();
-                    return returnType;
-                } else {
-                    throw createKafkaError("invalid " + functionName + " remote function signature");
-                }
+                Type returnType = function.getReturnType();
+                return returnType;
             }
         }
         return function;